requirements.txt
django
djangorestframework
python-dateutil